refactor enums
