+++ src/main/java/BuenSabor/enums/FormaPago.java
@@ -1,4 +1,4 @@
-package BuenSabor.model;
+package BuenSabor.enums;
 
 public enum FormaPago {
     Efectivo,
